agent-webapp/requirements.txt
# Agent 365 Web Service
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0

# Azure Identity & Foundry
azure-identity
requests
python-dotenv
